wasm: add Emscripten/WASM compilation support
- platform.h: Add __EMSCRIPTEN__ detection, QWordSwapC function, VALVE_LITTLE_ENDIAN
- fasttimer.h: Use emscripten_get_now() for CCycleCount::Sample()
- mathlib.h: Replace x86 inline asm (fsincos, fistpl) with portable C equivalents
- utlblockmemory.h: Fix bitfield std::swap workaround for Emscripten
- diff.cpp, strtools.cpp, RichText.cpp: Fix min() type mismatch with explicit casts
- reliabletimer.cpp: Use emscripten_get_now() for timing
- processor_detect_linux.cpp: Stub CPU detection for WASM
- pathmatch.cpp: Use __builtin_trap() for DEBUG_BREAK on WASM
- CMakeLists.txt: Add -msimd128 -D__SSE__ -D__SSE2__ flags for Emscripten

All static libraries now compile to WASM: tier1, tier2, vgui_controls

+++ tier1/strtools.cpp
@@ -2005,7 +2005,7 @@ bool V_ExtractFilePath (const char *path, char *dest, int destSize )
 		src--;
 	}
 
-	int copysize = min( src - path, destSize - 1 );
+	int copysize = min( (int)(src - path), destSize - 1 );
 	memcpy( dest, path, copysize );
 	dest[copysize] = 0;
